- more consistent API between FileGraph and LC_*Graph  - implement missing functions in FileGraph needed by LC_*Graph
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@8712 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ include/Galois/Graphs/Graph.h
@@ -561,7 +561,7 @@ public:
   //// General Things ////
 
   //! Returns the number of neighbors
-  int neighborsSize(GraphNode N, Galois::MethodFlag mflag = ALL) const {
+  size_t neighborsSize(GraphNode N, Galois::MethodFlag mflag = ALL) const {
     assert(N.ID);
     acquire(N.ID, mflag);
     return N.ID->edges.size();
